tests/requirements/base.txt
pytest
pytest-django
pytest-async
pytest-cov
pytest-factoryboy
coverage
coveralls>=2
django-taggit-helpers
djangocms-text
djangocms-video
pysolr
-e .
